requirements.txt
faiss-cpu
sentence-transformers
markdown
tqdm
numpy
